Added function for path averaging.
Changed A-->B so output is the same. A for loop will calculate the average of the first n entries in B and put this into a new array A. This is continued on a range one higher for the next entry and so on to create a full array A containing path averaged values of B. B is a pointer so care needs to be taken.